Refactor new tests to conform to best practices.
