+++ tests/Aquifer.API.IntegrationTests/Endpoints/Bibles/Books/List/EndpointTests.cs
@@ -3,14 +3,14 @@ using System.Net;
 using Aquifer.API.Endpoints.BibleBooks.List;
 using FastEndpoints.Testing;
 
-namespace Aquifer.API.IntegrationTests;
+namespace Aquifer.API.IntegrationTests.Endpoints.Bibles.Books.List;
 
-public sealed class BibleBooksListTestFixture(AppFixture _appFixture) : TestBase<AppFixture>
+public sealed class EndpointTests(App _app) : TestBase<App>
 {
     [Fact]
     public async Task ValidRequest_ShouldReturnSuccess()
     {
-        var (response, result) = await _appFixture.Client.GETAsync<Endpoint, IReadOnlyList<Response>>();
+        var (response, result) = await _app.Client.GETAsync<Endpoint, IReadOnlyList<Response>>();
 
         response.StatusCode.Should().Be(HttpStatusCode.OK);
         result.Should().HaveCountGreaterThan(80);
